expose number of lost packets without passing through an error (#735)

+++ pkg/rtplossdetector/lossdetector.go
@@ -19,5 +19,5 @@ func New() *LossDetector {
 // Process processes a RTP packet.
 // It returns the number of lost packets.
 func (r *LossDetector) Process(pkt *rtp.Packet) uint {
-	return (*rtplossdetector.LossDetector)(r).Process(pkt)
+	return uint((*rtplossdetector.LossDetector)(r).Process(pkt))
 }
